moved helper function from controller into a service
